+++ commands/events.ts
@@ -5,7 +5,7 @@ import HellHub, { type GlobalEvent } from "@hellhub-collective/sdk";
 
 import ascii from "utils/ascii";
 import request from "utils/request";
-import { createListCommand, parseListOptions } from "utils/list-options";
+import { createListCommand, parseListOptions } from "utils/options";
 
 const border = (message: string) => {
   if (message.includes("MAJOR")) {
