+++ commands/stratagems.ts
@@ -5,7 +5,7 @@ import HellHub, { type Stratagem } from "@hellhub-collective/sdk";
 
 import ascii from "utils/ascii";
 import request from "utils/request";
-import { createListCommand, parseListOptions } from "utils/list-options";
+import { createListCommand, parseListOptions } from "utils/options";
 
 export default function stratagems(program: Command) {
   createListCommand(
